* 'master' of git://github.com/gitextensions/gitextensions: (687 commits)
  File History fixed if Settings.RevisionGraphShowWorkingDirChanges is on
  GetDiffFiles, GetStashDiffFiles and GetTreeFiles return IList instead of IEnumerable
  You break my heart (or code)
  see d958d61ecd071: use ToList() only when it is needed
  Two tests fixed
  "0 hours" text fixed
  Git working dir fixed for shell extension and VS addin after last commits
  Remove submodule command fixed
  Exception fixed for first revision in repository (Diff tab) after commit c982a8127d589
  FindGitWorkingDir should be called explicitly before creating GitModule
  Suggested local branch name fixed for svn remote branches
  Git SVN Clone support custom trunk, tags and branches directory
  Auto pull option behavior improved
  Bug fixed in ProcessSubmodulePatch. Reset form can delete directories
  Revert "IconDiffTool removed"
  GitEx shell extension registration in settings form fixed with UAC
  Icons for GitEx shell context menu added. Closes #710
  IconDiffTool removed
  Load heads async
  Revert "Check for null. Removed reverse - there is already ByCommitDateDescending"
  ...

+++ NetSpell.SpellChecker/Dictionary/Affix/AffixUtility.cs
@@ -31,7 +31,7 @@ namespace NetSpell.SpellChecker.Dictionary.Affix
 					int passCount = 0;
 					for (int i = 0;  i < entry.ConditionCount; i++) 
 					{
-						int charCode = (int)word[i];
+						int charCode = word[i];
 						if ((entry.Condition[charCode] & (1 << i)) == (1 << i))
 						{
 							passCount++;
@@ -78,7 +78,7 @@ namespace NetSpell.SpellChecker.Dictionary.Affix
 					int passCount = 0;
 					for (int i = 0;  i < entry.ConditionCount; i++) 
 					{
-						int charCode = (int)word[word.Length - (entry.ConditionCount - i)];
+						int charCode = word[word.Length - (entry.ConditionCount - i)];
 						if ((entry.Condition[charCode] & (1 << i)) == (1 << i))
 						{
 							passCount++;
@@ -177,7 +177,7 @@ namespace NetSpell.SpellChecker.Dictionary.Affix
 							// turn off chars in member group
 							for (int j=0; j < numMember; j++) 
 							{
-								int charCode = (int)memberChars[j];
+								int charCode = memberChars[j];
 								entry.Condition[charCode] = entry.Condition[charCode] & ~(1 << num);
 							}
 						}
@@ -186,7 +186,7 @@ namespace NetSpell.SpellChecker.Dictionary.Affix
 							// turn on chars in member group
 							for (int j=0; j < numMember; j++) 
 							{
-								int charCode = (int)memberChars[j];
+								int charCode = memberChars[j];
 								entry.Condition[charCode] = entry.Condition[charCode] | (1 << num);
 							}
 						}
@@ -207,7 +207,7 @@ namespace NetSpell.SpellChecker.Dictionary.Affix
 						else 
 						{
 							// turn on char
-							int charCode = (int)cond;
+							int charCode = cond;
 							entry.Condition[charCode] = entry.Condition[charCode] | (1 << num);
 						}
 					} // not group
@@ -257,7 +257,7 @@ namespace NetSpell.SpellChecker.Dictionary.Affix
 				int passCount = 0;
 				for (int i = 0;  i < entry.ConditionCount; i++) 
 				{
-					int charCode = (int)tempWord[i];
+					int charCode = tempWord[i];
 					if ((entry.Condition[charCode] & (1 << i)) == (1 << i))
 					{
 						passCount++;
@@ -305,7 +305,7 @@ namespace NetSpell.SpellChecker.Dictionary.Affix
 				int passCount = 0;
 				for (int i = 0;  i < entry.ConditionCount; i++) 
 				{
-					int charCode = (int)tempWord[tempWord.Length - (entry.ConditionCount - i)];
+					int charCode = tempWord[tempWord.Length - (entry.ConditionCount - i)];
 					if ((entry.Condition[charCode] & (1 << i)) == (1 << i))
 					{
 						passCount++;
